Swap primary and secondary colors
It's a bit strange to fit the existing designs into a MUI theme because
MUI uses the same color for buttons and the app bar, but in the designs,
those are different colors. I think it'll be easier to set the primary
as the button color and just override the default props when necessary.