Enable OSSF scorecard workflow
Start using the [OpenSSF scorecard](https://github.com/ossf/scorecard) to track overall health
and security of this repository.  This PR also adds a badge to the main page to see the results,
which can be clicked on to see details (which will only show results once the PR is merged
causing the workflow to run for the first time).

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>